+++ packages/block-editor/src/components/colors/index.js
@@ -3,7 +3,5 @@ export {
 	getColorObjectByAttributeValues,
 	getColorObjectByColorValue,
 } from './utils';
-export {
-	createCustomColorsHOC,
-	default as withColors,
-} from './with-colors';
+export { default as useColors, useCustomColors } from './use-colors';
+export { createCustomColorsHOC, default as withColors } from './with-colors';