+++ packages/block-editor/src/components/colors/with-colors.js
@@ -1,148 +1,12 @@
 /**
- * External dependencies
- */
-import { get, isString, kebabCase, reduce, upperFirst } from 'lodash';
-
-/**
  * WordPress dependencies
  */
-import { Component } from '@wordpress/element';
-import { withSelect } from '@wordpress/data';
-import { compose, createHigherOrderComponent } from '@wordpress/compose';
+import { createHigherOrderComponentWithMergeProps } from '@wordpress/compose';
 
 /**
  * Internal dependencies
  */
-import { getColorClassName, getColorObjectByColorValue, getColorObjectByAttributeValues, getMostReadableColor } from './utils';
-
-const DEFAULT_COLORS = [];
-
-/**
- * Higher order component factory for injecting the `colorsArray` argument as
- * the colors prop in the `withCustomColors` HOC.
- *
- * @param {Array} colorsArray An array of color objects.
- *
- * @return {function} The higher order component.
- */
-const withCustomColorPalette = ( colorsArray ) => createHigherOrderComponent( ( WrappedComponent ) => ( props ) => (
-	<WrappedComponent { ...props } colors={ colorsArray } />
-), 'withCustomColorPalette' );
-
-/**
- * Higher order component factory for injecting the editor colors as the
- * `colors` prop in the `withColors` HOC.
- *
- * @return {function} The higher order component.
- */
-const withEditorColorPalette = () => withSelect( ( select ) => {
-	const settings = select( 'core/block-editor' ).getSettings();
-	return {
-		colors: get( settings, [ 'colors' ], DEFAULT_COLORS ),
-	};
-} );
-
-/**
- * Helper function used with `createHigherOrderComponent` to create
- * higher order components for managing color logic.
- *
- * @param {Array}    colorTypes       An array of color types (e.g. 'backgroundColor, borderColor).
- * @param {Function} withColorPalette A HOC for injecting the 'colors' prop into the WrappedComponent.
- *
- * @return {Component} The component that can be used as a HOC.
- */
-function createColorHOC( colorTypes, withColorPalette ) {
-	const colorMap = reduce( colorTypes, ( colorObject, colorType ) => {
-		return {
-			...colorObject,
-			...( isString( colorType ) ? { [ colorType ]: kebabCase( colorType ) } : colorType ),
-		};
-	}, {} );
-
-	return compose( [
-		withColorPalette,
-		( WrappedComponent ) => {
-			return class extends Component {
-				constructor( props ) {
-					super( props );
-
-					this.setters = this.createSetters();
-					this.colorUtils = {
-						getMostReadableColor: this.getMostReadableColor.bind( this ),
-					};
-
-					this.state = {};
-				}
-
-				getMostReadableColor( colorValue ) {
-					const { colors } = this.props;
-					return getMostReadableColor( colors, colorValue );
-				}
-
-				createSetters() {
-					return reduce( colorMap, ( settersAccumulator, colorContext, colorAttributeName ) => {
-						const upperFirstColorAttributeName = upperFirst( colorAttributeName );
-						const customColorAttributeName = `custom${ upperFirstColorAttributeName }`;
-						settersAccumulator[ `set${ upperFirstColorAttributeName }` ] =
-							this.createSetColor( colorAttributeName, customColorAttributeName );
-						return settersAccumulator;
-					}, {} );
-				}
-
-				createSetColor( colorAttributeName, customColorAttributeName ) {
-					return ( colorValue ) => {
-						const colorObject = getColorObjectByColorValue( this.props.colors, colorValue );
-						this.props.setAttributes( {
-							[ colorAttributeName ]: colorObject && colorObject.slug ? colorObject.slug : undefined,
-							[ customColorAttributeName ]: colorObject && colorObject.slug ? undefined : colorValue,
-						} );
-					};
-				}
-
-				static getDerivedStateFromProps( { attributes, colors }, previousState ) {
-					return reduce( colorMap, ( newState, colorContext, colorAttributeName ) => {
-						const colorObject = getColorObjectByAttributeValues(
-							colors,
-							attributes[ colorAttributeName ],
-							attributes[ `custom${ upperFirst( colorAttributeName ) }` ],
-						);
-
-						const previousColorObject = previousState[ colorAttributeName ];
-						const previousColor = get( previousColorObject, [ 'color' ] );
-						/**
-						* The "and previousColorObject" condition checks that a previous color object was already computed.
-						* At the start previousColorObject and colorValue are both equal to undefined
-						* bus as previousColorObject does not exist we should compute the object.
-						*/
-						if ( previousColor === colorObject.color && previousColorObject ) {
-							newState[ colorAttributeName ] = previousColorObject;
-						} else {
-							newState[ colorAttributeName ] = {
-								...colorObject,
-								class: getColorClassName( colorContext, colorObject.slug ),
-							};
-						}
-						return newState;
-					}, {} );
-				}
-
-				render() {
-					return (
-						<WrappedComponent
-							{ ...{
-								...this.props,
-								colors: undefined,
-								...this.state,
-								...this.setters,
-								colorUtils: this.colorUtils,
-							} }
-						/>
-					);
-				}
-			};
-		},
-	] );
-}
+import useColors, { useCustomColors } from './use-colors';
 
 /**
  * A higher-order component factory for creating a 'withCustomColors' HOC, which handles color logic
@@ -167,10 +31,11 @@ function createColorHOC( colorTypes, withColorPalette ) {
  * @return {Function} Higher-order component.
  */
 export function createCustomColorsHOC( colorsArray ) {
-	return ( ...colorTypes ) => {
-		const withColorPalette = withCustomColorPalette( colorsArray );
-		return createHigherOrderComponent( createColorHOC( colorTypes, withColorPalette ), 'withCustomColors' );
-	};
+	return ( ...colorTypes ) =>
+		createHigherOrderComponentWithMergeProps(
+			() => useCustomColors( colorTypes, colorsArray ),
+			'withCustomColors'
+		);
 }
 
 /**
@@ -198,6 +63,8 @@ export function createCustomColorsHOC( colorsArray ) {
  * @return {Function} Higher-order component.
  */
 export default function withColors( ...colorTypes ) {
-	const withColorPalette = withEditorColorPalette();
-	return createHigherOrderComponent( createColorHOC( colorTypes, withColorPalette ), 'withColors' );
+	return createHigherOrderComponentWithMergeProps(
+		() => useColors( colorTypes ),
+		'withColors'
+	);
 }